requirements.txt
asgiref==3.2.7
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-on-heroku==1.0.1
django-registration==3.1
django-sendgrid-v5==0.8.1
future==0.18.2
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
python-decouple==3.3
python-http-client==3.2.7
pytz==2020.1
sendgrid==6.3.1
sqlparse==0.3.1
whitenoise==5.1.0

